requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv
google-generativeai